Exclude `\s*$` from tokenised matches
Some implementations of TMCG (TextMate-compatible grammars); i.e., Atom,
will include the newline after `$` but before `^` as part of the initial
line's match. This prevents the next line from being tokenised correctly
when start-of-line patterns are involved (i.e., those that match `^`).

This behaviour is a bug with First Mate (Atom's implementation of TMCGs)
but is beyond fixing at this point...